==> 5_did_additionalrestriction.qmd <==
```{r echo = FALSE, include=FALSE}
library(tidyverse)
library(survey)
library(ggpubr)
library(modelsummary)
library(haven)

here::i_am("5_did_additionalrestriction.qmd")
options(box.path = here::here())
box::use(./functions)
box::reload(functions)
options(survey.lonely.psu="adjust")

df_intensive <- readRDS("data/df_intensive.rds")
df_extensive <- readRDS("data/df_extensive.rds")
```


# Additional sample restriction

```{r echo=FALSE}
df_intensive <- df_intensive |> 
  filter(age %in% 21:30 & educ %in% c("3", "4")) |>
  filter(decil <= 8) |>
  filter(urb == 1) |> 
  filter(!region %in% c(1, 2, 12)) |> 
  filter(civilstatus %in% c(2,5))

df_extensive <- df_extensive |> 
  filter(age %in% 21:30 & educ %in% c("3", "4")) |>
  filter(decil <= 8) |>
  filter(urb == 1) |> 
  filter(!region %in% c(1, 2, 12)) |> 
  filter(civilstatus %in% c(2,5))

intensive_design <- svydesign(
  id = ~varunit, 
  strata = ~varstrat, 
  weights = ~expr, 
  data = df_intensive, 
  nest = TRUE
)

extensive_design <- svydesign(
  id = ~varunit, 
  strata = ~varstrat, 
  weights = ~expr, 
  data = df_extensive, 
  nest = TRUE,
)
```

# Mean plots 

```{r}
# Intensive margin
meanplot_jobincome <- functions$plot_mean(df_intensive, "log_jobincome", title = "Log Job Income") 
meanplot_totalincome <- functions$plot_mean(df_intensive, "log_totalincome", title = "Log Total Income")
meanplot_hours <- functions$plot_mean(df_intensive, "hours", title = "Working Hours")

# Extensive margin 
meanplot_lfp <- functions$plot_mean(df_extensive, "lfp", title = "Labor Force Participation")
meanplot_employment <- functions$plot_mean(df_extensive, "employment", title = "Employment")
meanplot_unemployment <- functions$plot_mean(df_extensive, "unemployment", title = "Unemployment")
meanplot_contract <- functions$plot_mean(df_extensive, "contract", title = "Contract")

# Arrange the plots into a 2x4 grid
meanplot <- ggarrange(
  meanplot_jobincome, meanplot_totalincome, meanplot_hours, 
  meanplot_lfp, meanplot_employment, meanplot_unemployment, meanplot_contract,
  ncol=2, nrow=4, common.legend = TRUE, legend="bottom"
) 

# Export the combined figure to a file
ggsave("meanplot_restricted.png", meanplot, width = 10, height = 13)
```

# Models

```{r}
# Intensive margin (no control)
mod_jobincome <- functions$run_ols_2(intensive_design, "log_jobincome", F, F)
mod_totalincome <- functions$run_ols_2(intensive_design, "log_totalincome", F, F)
mod_hours <- functions$run_ols_2(intensive_design, "hours", F, F)

# Intensive margin (control)
mod_jobincome_control <- functions$run_ols_2(intensive_design, "log_jobincome", T, F)
mod_totalincome_control <- functions$run_ols_2(intensive_design, "log_totalincome", T, F)
mod_hours_control <- functions$run_ols_2(intensive_design, "hours", T, F)


# Extensive margin (no control)
mod_lfp <- functions$run_ols_2(extensive_design, "lfp", F, F)
mod_employment <- functions$run_ols_2(extensive_design, "employment", F, F)
mod_unemployment <- functions$run_ols_2(extensive_design, "unemployment", F, F)
mod_contract <- functions$run_ols_2(extensive_design, "contract", F, F)

# Extensive margin (control)
mod_lfp_control <- functions$run_ols_2(extensive_design, "lfp", T, F)
mod_employment_control <- functions$run_ols_2(extensive_design, "employment", T, F)
mod_unemployment_control <- functions$run_ols_2(extensive_design, "unemployment", T, F)
mod_contract_control <- functions$run_ols_2(extensive_design, "contract", T, F)
```


```{r}
regtable_intensive <- modelsummary(
  list(
    mod_jobincome, mod_jobincome_control,
    mod_totalincome, mod_totalincome_control,
    mod_hours, mod_hours_control
  ),
  coef_omit = "^(?!eligible:period)",
  gof_map = c("nobs", "r.squared"),
  output = "latex",
  stars = TRUE
)

regtable_extensive <- modelsummary(
  list(
    mod_lfp, mod_lfp_control,
    mod_employment, mod_employment_control,
    mod_unemployment, mod_unemployment_control, 
    mod_contract, mod_contract_control
  ),
  coef_omit = "^(?!eligible:period)",
  gof_map = c("nobs", "r.squared"),
  output = "latex",
  stars = TRUE
)
```

# Event study plot (with controls)

```{r}
# Intensive
eventstudyplot_jobincome <- functions$plot_event_study(mod_jobincome_control, output_title = "Log Job Income")
eventstudyplot_totalincome <- functions$plot_event_study(mod_totalincome_control, output_title = "Log Total Income")
eventstudyplot_hours <- functions$plot_event_study(mod_hours_control, output_title = "Working Hours")

# Extensive  
eventstudyplot_lfp <- functions$plot_event_study(mod_lfp_control, output_title = "Labor Force Participation")
eventstudyplot_employment <- functions$plot_event_study(mod_employment_control, output_title = "Employment")
eventstudyplot_unemployment <- functions$plot_event_study(mod_unemployment_control, output_title = "Unemployment")
eventstudyplot_contract <- functions$plot_event_study(mod_contract_control, output_title = "Contract")

# Plot
eventstudyplot <- ggarrange(
  eventstudyplot_jobincome, eventstudyplot_totalincome, eventstudyplot_hours, 
  eventstudyplot_lfp, eventstudyplot_employment, eventstudyplot_unemployment, eventstudyplot_contract, 
  ncol=2, nrow=4
) 

ggsave("eventstudyplot_restricted.png", eventstudyplot, width = 8, height = 12)
```
